requirements.txt
accelerate==1.6.0
acres==0.3.0
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
docx==0.2.4
etelemetry==0.3.1
fastapi==0.115.12
filelock==3.18.0
fitz==0.0.1.dev2
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gguf==0.14.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.23.5
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isodate==0.6.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.23
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langsmith==0.3.30
llama_cpp_python==0.3.8
looseversion==1.3.0
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==1.26.4
openai==1.55.3
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
pluggy==1.5.0
portalocker==2.10.1
prometheus_client==0.20.0
propcache==0.3.1
protobuf==4.25.6
prov==2.0.1
psutil==7.0.0
puremagic==1.28
pydantic==2.11.3
pydantic_core==2.33.1
pydot==3.0.4
Pygments==2.19.1
PyMuPDF==1.25.5
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
qdrant-client==1.13.3
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==4.0.2
sentencepiece==0.2.0
setuptools==78.1.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.20.3
torch==2.2.2
tqdm==4.67.1
traits==7.0.2
transformers==4.46.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
wget==3.2
yarl==1.19.0
zstandard==0.23.0
